fix: bug s3-download with 1 key as a prefix
When 1 key is specified for "s3-download" maven-task, this key must be
only key for file. On the other side, when specified 2 or more keys, all
they are processed as prefixes. With this commit prefixes are ands with
"/". Prefix can be specialized on every place in "keys" parameters or
quite not specialized.

Signed-off-by: Vladimir V. Bychkov <github@bychkov.name>